created single post component and styled. Moved all components into components folder instead of just out in the wild
