+++ src/App.js
@@ -1,5 +1,6 @@
-import TopBar from './topbar/Topbar';
+import TopBar from './components/topbar/Topbar';
 import './App.css';
+import Single from './pages/single/Single'
 
 import Home from './pages/home/Home';
 
@@ -7,7 +8,7 @@ function App() {
   return (
     <>
     <TopBar />
-    <Home />
+    <Single />
     </>
    
   );
