+++ src/components/singlePost/singlepost.css
@@ -1,0 +1,59 @@
+
+.singlePost {
+    flex: 9;
+}
+
+.singlePostWrapper {
+    padding: 20px;
+    padding-right: 0;
+}
+
+.singlePostImg{
+    width: 100%;
+    height: 300px;
+    border-radius: 5px;
+    object-fit: cover;
+}
+
+.singlePostTitle {
+    text-align: center;
+    margin: 10px;
+    font-size: 28px;
+    font-family: sans-serif;
+}
+
+.singlePostEdit {
+    float: right;
+    font-size: 16px;
+}
+.singlePostIcon {
+    margin-left: 10px;
+    cursor: pointer;
+}
+
+.singlePostIcon:first-child{
+    color: teal;
+}
+.singlePostIcon:last-child{
+    color: tomato;
+}
+.singlePostInfo {
+    margin-bottom: 20px;
+    display: flex;
+    justify-content: space-between;
+    font-size: 16px;
+    color: #b39656;
+    font-family: sans-serif;
+
+}
+
+.singlePostDescription{
+    color: #444;
+    font-size: 20px;
+    line-height: 25px;
+}
+.singlePostDescription::first-letter {
+    margin-left: 20px;
+    font-size: 30px;
+    font-weight: 600;
+}
