+++ src/pages/home/Home.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import './Home.css'
 
-import Header from '../../header/Header'
-import Posts from '../../posts/Posts'
-import SideBar from '../../sidebar/Sidebar'
+import Header from '../../components/header/Header'
+import Posts from '../../components/posts/Posts'
+import SideBar from '../../components/sidebar/Sidebar'
 
 export default function Home () {
     return (
