+++ src/pages/single/single.css
@@ -1,0 +1,4 @@
+.single{
+    display: flex;
+}
+
